add search function, and refactor templates
