+++ src/main/java/com/example/bookstore/dao/BooksRepository.java
@@ -3,5 +3,9 @@ package com.example.bookstore.dao;
 import com.example.bookstore.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface BooksRepository extends JpaRepository<Book, Integer> {
+
+    List<Book> findByTitleContainingIgnoreCase(String title);
 }
