+++ src/main/resources/static/css/home.css
@@ -1,110 +1,5 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-}
-
-header {
-    background-color: #f8f8f8;
-    border-bottom: 1px solid #ccc;
-}
-
-.contact-info {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px;
-    background-color: #fff;
-    font-size: 14px;
-    color: #444;
-}
-
-.contact-info .right {
-    color: #888;
-}
-
-.logo-search-cart {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 10px 20px;
-    background-color: #f8f8f8;
-}
-
-.logo h1 {
-    margin: 0;
-    color: #ff6a00;
-    font-size: 28px;
-}
-
-.search-bar {
-    flex-grow: 2;
-    display: flex;
-    align-items: center;
-}
-
-.search-bar select, .search-bar input, .search-bar button {
-    padding: 8px;
-    margin-right: 5px;
-    font-size: 14px;
-}
-
-.search-bar input {
-    flex-grow: 1;
-}
-
-.cart {
-    display: flex;
-    align-items: center;
-    color: #333;
-}
-
-.cart span {
-    margin-left: 10px;
-}
-
-nav {
-    background-color: #ff6a00;
-    color: white;
-}
-
-nav ul {
-    display: flex;
-    justify-content: center;
-    list-style-type: none;
-    margin: 0;
-    padding: 10px 0;
-}
-
-nav ul li {
-    margin: 0 15px;
-    cursor: pointer;
-}
-
 .container {
     display: flex;
-}
-
-.sidebar {
-    width: 250px;
-    background-color: #f8f8f8;
-    border-right: 1px solid #ccc;
-    padding: 20px;
-}
-
-.sidebar ul {
-    list-style-type: none;
-    padding: 0;
-}
-
-.sidebar ul li {
-    padding: 10px 0;
-    cursor: pointer;
-    border-bottom: 1px solid #ccc;
-}
-
-.sidebar ul li:hover {
-    background-color: #ff6a00;
-    color: white;
 }
 
 .content {
@@ -210,58 +105,7 @@ nav ul li {
     border-radius: 8px;
 }
 
-footer {
-    background-color: #f8f8f8;
-    padding: 20px;
-    font-size: 14px;
-    border-top: 1px solid #ccc;
-}
-
-.footer-container {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-}
-
-.footer-section {
-    width: 22%;
-    padding: 10px;
-}
-
-.footer-section h4 {
-    color: black;
-    font-size: 18px;
-    margin-bottom: 10px;
-}
-
-.footer-section p, .footer-section ul {
-    color: #333;
-    margin: 0;
-    padding: 0;
-}
-
-.footer-section ul {
-    list-style-type: none;
-}
-
-.footer-section ul li {
-    margin-bottom: 5px;
-}
-
-.footer-bottom {
-    text-align: center;
-    padding: 10px 0;
-    border-top: 1px solid #ccc;
-    color: #888;
-}
-
-.social-icons {
-    display: flex;
-    gap: 10px;
-}
-
 .coming-soon {
     color: blue;
     font-weight: bold;
 }
-
