+++ src/main/resources/templates/layout.html
@@ -1,0 +1,138 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/css/layout.css" />
+</head>
+
+<body>
+<!-- Header Fragment -->
+<header th:fragment="header">
+    <div class="contact-info">
+        <span>Liên hệ: <strong>0708-744-001</strong> (Call/Zalo)</span>
+        <span class="right">Hướng dẫn đặt sách | Theo dõi đơn hàng | Tài Khoản</span>
+    </div>
+    <div class="logo-search-cart">
+        <div class="logo">
+            <h1>JBooks</h1>
+        </div>
+        <div class="search-bar">
+            <form th:action="@{/home}" method="get">
+                <input type="text" name="search" placeholder="Tìm sách tại đây" th:value="${search}" />
+                <button type="submit">Tìm</button>
+            </form>
+        </div>
+        <div class="cart">
+            <span>Giỏ Hàng</span>
+            <span>0đ</span>
+        </div>
+    </div>
+    <nav>
+        <ul>
+            <li>Home</li>
+            <li>Sách học</li>
+            <li>Sách luyện thi</li>
+            <li>Sách tham khảo</li>
+            <li>Flash Card</li>
+            <li>Học Online</li>
+            <li>Flash Sale</li>
+        </ul>
+    </nav>
+</header>
+
+<div class="container">
+    <!-- Sidebar Fragment -->
+    <aside class="sidebar" th:fragment="sidebar">
+        <ul>
+            <li>Sách vở lòng</li>
+            <li>Minna no Nihongo – Bản mới</li>
+            <li>Minna no Nihongo – Bản cũ</li>
+            <li>Minna no Nihongo – XB Trẻ</li>
+            <li>Minna Trung Cấp</li>
+            <li>Luyện thi N5</li>
+            <li>Luyện thi N4</li>
+            <li>Luyện thi N3</li>
+            <li>Luyện thi N2</li>
+            <li>Luyện thi N1</li>
+            <li>Từ điển</li>
+            <li>Sách tham khảo</li>
+            <li>Sách chuyên ngành</li>
+            <li>Flash Card</li>
+            <li>Combo sách</li>
+        </ul>
+    </aside>
+
+    <!-- Main Content Area -->
+    <main layout:fragment="content">
+        <!-- Default content of the page -->
+    </main>
+</div>
+
+<!-- Footer Fragment -->
+<footer th:fragment="footer">
+    <div class="footer-container">
+        <div class="footer-section">
+            <h4>Liên hệ chúng tôi:</h4>
+            <p>
+                HOTLINE: <span style="color: green">0708-744-001</span> (8:00 –
+                18:00)
+            </p>
+            <p>
+                <strong>Sách tiếng Nhật JBooks</strong><br />
+                Số 1 Đại Cồ Việt, Bách Khoa, Hai Bà Trưng, Hà Nội<br />
+                <a href="mailto:vnjpbook@gmail.com" style="color: red">javbooks@gmail.com</a>
+            </p>
+            <p>
+                Có thể liên hệ qua zalo: 0123-666-999<br />
+                Các bạn ở Hà Nội nếu có nhu cầu mua sách tận nơi thì liên hệ
+                0987-999-666.<br />
+                Shop sẽ chiết khấu dành riêng cho các bạn ghé mua tại nơi nhé!
+            </p>
+        </div>
+        <div class="footer-section">
+            <h4>Về chúng tôi:</h4>
+            <p>
+                Với mong muốn gây dựng một cộng đồng học tiếng Nhật, giúp đỡ và chia
+                sẻ những kiến thức và kinh nghiệm. Năm 2010 JAVBOOK được ra đời nhằm
+                giúp các học viên nâng cao năng lực Nhật ngữ mà không mất chi phí.
+            </p>
+            <p>
+                Tiếp nối sự thành công của JAVBOOK, JBOOKS được ra đời để cung cấp
+                tài liệu cho các bạn học viên, giáo viên và nhà nghiên cứu giảng dạy
+                tiếng Nhật. Team đang cố gắng dịch, biên tập tài liệu giúp các bạn
+                có nhiều tài liệu để tham khảo hơn. Cám ơn các bạn đã ủng hộ!
+            </p>
+        </div>
+        <div class="footer-section">
+            <h4>Chính sách và hướng dẫn:</h4>
+            <ul>
+                <li>Hướng dẫn đặt sách</li>
+                <li>CD và Download file</li>
+                <li>Mã giảm giá</li>
+                <li>Hủy đơn hàng</li>
+                <li>Chính sách đổi trả</li>
+                <li>Chính sách vận chuyển</li>
+                <li>Chính sách bảo mật</li>
+            </ul>
+            <h4>Kết nối với chúng tôi:</h4>
+            <div class="social-icons">
+                <!-- Icons or links to social media -->
+            </div>
+        </div>
+        <div class="footer-section">
+            <h4>Tải ứng dụng:</h4>
+            <ul>
+                <li>Học tiếng Nhật JBOOKS</li>
+                <li>JBOOKS Audio <span class="coming-soon">COMING SOON!</span></li>
+            </ul>
+            <!-- Placeholder for QR codes or other images -->
+        </div>
+    </div>
+    <div class="footer-bottom">
+        <p>JBOOKS - Sách tiếng Nhật dành cho người Việt</p>
+        <p>© 2022 JBOOKS</p>
+    </div>
+</footer>
+</body>
+</html>
